requirements.txt
flask==2.3.3
flask-cors==4.0.0
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
transformers==4.31.0
torch==2.0.1
python-dotenv==1.0.0
pymongo==4.5.0
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0 